Updated to include helper scripts to verify success
