+++ e2e/utils/helper.ts
@@ -435,4 +435,33 @@ export const switchToFlow = async ({ page, extensionId }) => {
   await getCurrentAddress(page);
 };
 
+export const waitForTransaction = async ({ page, successtext = 'success' }) => {
+  // Wait for the transaction to be completed
+  await page.waitForURL(/.*dashboard\?activity=1.*/);
+  const url = await page.url();
+
+  const txId = url.match(/[\?&]txId=(\w+)/i)?.[1];
+
+  expect(txId).toBeDefined();
+
+  const progressBar = page.getByRole('progressbar');
+  await expect(progressBar).toBeVisible();
+  // Get the pending item with the cadence txId that was put in the url and status is pending
+  const pendingItem = page.getByTestId(new RegExp(`^.*${txId}.*$`)).filter({ hasText: 'Pending' });
+
+  await expect(pendingItem).toBeVisible({
+    timeout: 60_000,
+  });
+  await expect(progressBar).not.toBeVisible({ timeout: 60_000 });
+
+  // Get the executed item with the cadence txId that was put in the url and status is success
+  const executedItem = page
+    .getByTestId(new RegExp(`^.*${txId}.*$`))
+    .filter({ hasText: successtext });
+
+  await expect(executedItem).toBeVisible({
+    timeout: 60_000,
+  });
+};
+
 export const expect = test.expect;
